Fix header buttons not clickable
Closes #211
